Update on "Split DifferentiableGraphOp out of graph_executor.cpp"
DifferentiableGraphOp and the graph executor are two independent things.
Spliting them out will make future development more convenient.

This PR is purely copy and paste, and could be easily reviewed.

In the future I will further refactor things in backward.cpp,
in order to provide Python API to visualize backward graph created by
autodiff (#28549).

[ghstack-poisoned]
