+++ caffe2/CMakeLists.txt
@@ -363,7 +363,7 @@ if (NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
     ${TORCH_SRC_DIR}/csrc/jit/pickler.cpp
     ${TORCH_SRC_DIR}/csrc/jit/unpickler.cpp
     ${TORCH_SRC_DIR}/csrc/jit/graph_executor.cpp
-    ${TORCH_SRC_DIR}/csrc/jit/backward.cpp
+    ${TORCH_SRC_DIR}/csrc/jit/grad_executor.cpp
     ${TORCH_SRC_DIR}/csrc/jit/import_source.cpp
     ${TORCH_SRC_DIR}/csrc/jit/import.cpp
     ${TORCH_SRC_DIR}/csrc/jit/pickle.cpp
