+++ tools/build_variables.py
@@ -85,6 +85,7 @@ libtorch_sources = [
     "torch/csrc/jit/pickler.cpp",
     "torch/csrc/jit/unpickler.cpp",
     "torch/csrc/jit/graph_executor.cpp",
+    "torch/csrc/jit/grad_executor.cpp",
     "torch/csrc/jit/import.cpp",
     "torch/csrc/jit/import_legacy.cpp",
     "torch/csrc/jit/pickle.cpp",
